Updated tests and passed
