+++ tests/Http/Controllers/Auth/LoginControllerTest.php
@@ -3,6 +3,7 @@
 namespace Tests\Http\Controllers\Auth;
 
 use Tests\TestCase;
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 
 class LoginControllerTest extends TestCase
@@ -31,7 +32,7 @@ class LoginControllerTest extends TestCase
     /** @test */
     public function login_authenticates_and_redirects_user()
     {
-        $user = \App\Models\User::factory()->create();
+        $user = User::factory()->create();
 
         $response = $this->post(route('login'), [
             'email' => $user->email,
