+++ tests/Http/Controllers/Auth/RegisterControllerTest.php
@@ -12,8 +12,6 @@ class RegisterControllerTest extends TestCase
     /** @test */
     public function user_is_created_when_they_register()
     {
-        $this->withoutExceptionHandling();
-
         $response = $this->postJson('/register', [
             'name' => 'Joseph Hand',
             'email' => 'joseph.hand@example.com',
@@ -28,17 +26,4 @@ class RegisterControllerTest extends TestCase
             'email' => 'joseph.hand@example.com'
         ]);
     }
-
-//    /** @test */
-//    public function only_an_authenticated_user_can_see_profile_view()
-//    {
-//        $this->getJson('profile')
-//             ->assertStatus(401);
-//
-//        $user = \App\Models\User::factory()->create();
-//
-//        $this->actingAs($user)
-//             ->getJson('home')
-//             ->assertOk();
-//    }
 }
